remove window following move to anemoi-utils
