+++ pyproject.toml
@@ -33,7 +33,7 @@ classifiers = [
 ]
 dynamic = [ "version" ]
 dependencies = [
-  "anemoi-utils>=0.5.1",
+  "anemoi-utils>=0.5.4",
   "cfunits",
   "earthkit-data>=0.12.4,<1",
   "earthkit-geo>=0.3",
