+++ src/anemoi/transform/filters/tabular/irregular_to_grid.py
@@ -13,11 +13,11 @@ from datetime import datetime
 import earthkit.data as ekd
 import numpy as np
 import pandas as pd
+from anemoi.utils.window import Window
 
 from anemoi.transform.filter import Filter
 from anemoi.transform.filters.tabular import filter_registry
 from anemoi.transform.filters.tabular.support.utils import raise_if_df_missing_cols
-from anemoi.transform.filters.tabular.support.window import Window
 
 LOG = logging.getLogger(__name__)
